Moving collectors into oaj.sparql.util.compose
