+++ jena-arq/src/test/java/org/apache/jena/sparql/util/compose/TestDatasetCollector.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.jena.query.util;
+package org.apache.jena.sparql.util.compose;
 
 import static org.apache.jena.rdf.model.ModelFactory.createModelForGraph;
 import static org.apache.jena.sparql.sse.SSE.parseGraph;
@@ -27,6 +27,7 @@ import org.apache.jena.graph.*;
 import org.apache.jena.query.Dataset;
 import org.apache.jena.query.DatasetFactory;
 import org.apache.jena.rdf.model.Model;
+import org.apache.jena.sparql.util.compose.DatasetCollector;
 import org.junit.Assert;
 import org.junit.Test;
